add  owner check

+++ controllers/contactsControllers.js
@@ -22,7 +22,8 @@ export const getAllContacts = async (req, res, next) => {
 export const getContactById = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const contact = await Contact.findById(id);
+    const { _id: owner } = req.user;
+    const contact = await Contact.findById(id).where("owner").equals(owner);
 
     if (!contact) {
       throw HttpError(404);
@@ -37,7 +38,8 @@ export const getContactById = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const deletedContact = await Contact.findByIdAndDelete(id);
+    const { _id: owner } = req.user;
+    const deletedContact = await Contact.findByIdAndDelete(id).where("owner").equals(owner);
 
     if (!deletedContact) {
       throw HttpError(404, "Not found");
@@ -75,10 +77,11 @@ export const updateContact = async (req, res, next) => {
     }
 
     const { id } = req.params;
-
+    const { _id: owner } = req.user;
+    
     const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
       new: true,
-    });
+    }).where("owner").equals(owner);
     if (!updatedContact) {
       throw HttpError(404, "Not found");
     }
@@ -101,10 +104,11 @@ export const updateContactFavoriteOption = async (req, res, next) => {
     }
 
     const { id } = req.params;
+    const { _id: owner } = req.user;
 
     const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
       new: true,
-    });
+    }).where("owner").equals(owner);
     if (!updatedContact) {
       throw HttpError(404, "Not found");
     }
